+++ src/EntityGraphQL/Compiler/MutationResult.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using System.Linq.Expressions;
 using EntityGraphQL.Schema;
@@ -6,11 +7,11 @@ namespace EntityGraphQL.Compiler
 {
     public class MutationResult : ExpressionResult
     {
-        private readonly Schema.MutationType mutationType;
+        private readonly MutationType mutationType;
         private readonly Expression paramExp;
         private readonly Dictionary<string, ExpressionResult> gqlRequestArgs;
 
-        public MutationResult(Schema.MutationType mutationType, Dictionary<string, ExpressionResult> args) : base(null)
+        public MutationResult(MutationType mutationType, Dictionary<string, ExpressionResult> args) : base(null)
         {
             this.mutationType = mutationType;
             this.gqlRequestArgs = args;
@@ -19,13 +20,13 @@ namespace EntityGraphQL.Compiler
 
         public override Expression Expression { get { return paramExp; } }
 
-        public object Execute(object[] externalArgs)
+        public object Execute(object contextArg, IServiceProvider serviceProvider)
         {
             try
             {
-                return mutationType.Call(externalArgs, gqlRequestArgs);
+                return mutationType.Call(contextArg, gqlRequestArgs, serviceProvider);
             }
-            catch(EntityQuerySchemaException e)
+            catch (EntityQuerySchemaException e)
             {
                 throw new EntityQuerySchemaException($"Error applying mutation: {e.Message}");
             }